fix exceptions for declare global opption

+++ gist/SwiftXrtBuild_2_rename.sh
@@ -16,7 +16,7 @@ for prod_ID in ${prod_IDs[@]}; do
     if [[ -d "USERPROD_${prod_ID}" ]]; then
         rm $spec_path -f &&
             ln -s $build_path/USERPROD_${prod_ID}/spec $spec_path
-    elif [[ x$(find ./ -name "*pc.pi" -printf "1") != x ]]; then
+    elif [[ x$(find ./ -regex ".*\(pc\|wt\)\.pi" -printf "1") != x ]]; then
         rm $spec_path -f &&
             ln -s $build_path $spec_path
     fi
@@ -37,42 +37,48 @@ for prod_ID in ${prod_IDs[@]}; do
     mkdir $spec_path/fit -p
 
     # for per Obs
-    obs_IDs=($(find . -name "Obs_*pc.pi" -printf "%f\n" |
-        sed -r -n "s/^\S*Obs_([0-9]+)pc\S*$/\1/p"))
+    obs_IDs=($(find . -name "Obs_*[pw][ct].pi" -printf "%f\n" |
+        sed -r -n "s/^\S*Obs_([0-9]+)(pc|wt)\S*$/\1/p"))
     for obs_ID in ${obs_IDs[@]}; do
         tmp_head=Obs_${obs_ID}
-        declare -A tmp_orig_names=(
-            ["nongrp"]=${tmp_head}pcsource.pi
-            ["grpauto"]=${tmp_head}pc.pi
-            ["bkg"]=${tmp_head}pcback.pi
-            ["rmf"]=${tmp_head}pc.rmf
-            ["arf"]=${tmp_head}pc.arf)
+        for cam in "pc" "wt"; do
+            declare -A tmp_orig_names=(
+                ["${cam}_nongrp"]=${tmp_head}${cam}source.pi
+                ["${cam}_grpauto"]=${tmp_head}${cam}.pi
+                ["${cam}_bkg"]=${tmp_head}${cam}back.pi
+                ["${cam}_rmf"]=${tmp_head}${cam}.rmf
+                ["${cam}_arf"]=${tmp_head}${cam}.arf)
 
-        for key in ${!tmp_orig_names[@]}; do
-            orig_name=${tmp_orig_names[$key]}
-            new_name=xrtBuild${prod_ID}_Obs${obs_ID}_${key}.fits
-            new_names[$key]=$new_name
-            cp -f $orig_name $spec_path/fit/$new_name
+            for key in ${!tmp_orig_names[@]}; do
+                orig_name=${tmp_orig_names[$key]}
+                if [[ ! -f "$orig_name" ]]; then continue; fi
+                new_name=xrtBuild${prod_ID}_Obs${obs_ID}_${key}.fits
+                new_names[$key]=$new_name
+                cp -f $orig_name $spec_path/fit/$new_name
+            done
         done
     done
 
     # for per project
-    proj_IDs=($(find . -name "[0-9]*pc.pi" -printf "%f\n" |
-        sed -r -n "s/^([0-9]+)pc\S*$/\1/p"))
+    proj_IDs=($(find . -name "[0-9]*[pw][ct].pi" -printf "%f\n" |
+        sed -r -n "s/^([0-9]+)(pc|wt)\S*$/\1/p"))
     for proj_ID in ${proj_IDs[@]}; do
         tmp_head=${proj_ID}
-        declare -A tmp_orig_names=(
-            ["nongrp"]=${tmp_head}pcsource.pi
-            ["grpauto"]=${tmp_head}pc.pi
-            ["bkg"]=${tmp_head}pcback.pi
-            ["rmf"]=${tmp_head}pc.rmf
-            ["arf"]=${tmp_head}pc.arf)
+        for cam in "pc" "wt"; do
+            declare -A tmp_orig_names=(
+                ["${cam}_nongrp"]=${tmp_head}${cam}source.pi
+                ["${cam}_grpauto"]=${tmp_head}${cam}.pi
+                ["${cam}_bkg"]=${tmp_head}${cam}back.pi
+                ["${cam}_rmf"]=${tmp_head}${cam}.rmf
+                ["${cam}_arf"]=${tmp_head}${cam}.arf)
 
-        for key in ${!tmp_orig_names[@]}; do
-            orig_name=${tmp_orig_names[$key]}
-            new_name=xrtBuild${prod_ID}_Proj${proj_ID}_${key}.fits
-            new_names[$key]=$new_name
-            cp -f $orig_name $spec_path/fit/$new_name
+            for key in ${!tmp_orig_names[@]}; do
+                orig_name=${tmp_orig_names[$key]}
+                if [[ ! -f "$orig_name" ]]; then continue; fi
+                new_name=xrtBuild${prod_ID}_Proj${proj_ID}_${key}.fits
+                new_names[$key]=$new_name
+                cp -f $orig_name $spec_path/fit/$new_name
+            done
         done
     done
 done
